requirements.txt
asgiref==3.2.7
boto==2.49.0
boto3==1.14.2
botocore==1.17.2
coverage==5.1
Django==3.0.7
django-storages==1.9.1
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
docutils==0.15.2
jmespath==0.10.0
Pillow==7.1.2
psycopg2==2.8.5
python-dateutil==2.8.1
python-dotenv==0.13.0
pytz==2020.1
s3transfer==0.3.3
six==1.15.0
sqlparse==0.3.1
urllib3==1.25.9
